Refactor and use extract_info_from_url
We are using the urlparse library to extract hostname information, but
it handles URLs without protocols (i.e. google.com/test vs
https://google.com/tesT) poorly.  The protocol-less string is parsed
entirely as a relative path, instead of identifying google.com as an
appropriate hostname.

This function (created for aiohttp) handles such cases in a better
manner, by forcing the parsing with a protocol if an initial parse
fails.
