+++ ddtrace/contrib/aiohttp/patch.py
@@ -19,6 +19,7 @@ from ...internal.schema import schematize_url_operation
 from ...pin import Pin
 from ...propagation.http import HTTPPropagator
 from ..trace_utils import ext_service
+from ..trace_utils import extract_info_from_url
 from ..trace_utils import set_http_meta
 from ..trace_utils import unwrap
 from ..trace_utils import with_traced_module as with_traced_module_sync
@@ -64,20 +65,6 @@ class _WrappedConnectorClass(wrapt.ObjectProxy):
             span.set_tag(COMPONENT, config.aiohttp.integration_name)
             result = await self.__wrapped__._create_connection(req, *args, **kwargs)
             return result
-
-
-def extract_info_from_url(url):
-    # type: (str) -> typing.Tuple[str, str]
-    parse_result = parse.urlparse(url)
-    query = parse_result.query
-
-    # Relative URLs don't have a netloc, so we force them
-    if not parse_result.netloc:
-        parse_result = parse.urlparse("//{url}".format(url=url))
-
-    netloc = parse_result.netloc.split("@", 1)[-1]  # Discard auth info
-    netloc = netloc.split(":", 1)[0]  # Discard port information
-    return netloc, query
 
 
 @with_traced_module
